EZ-2102: Fixed month input (2 digits int) and converted all money fields to decimals

+++ Backend/Strategies/PricingModel/PricingModelModel.cs
@@ -3,7 +3,7 @@
 	public class PricingModelModel
 	{
 		// Main input fields
-		public int LoanAmount { get; set; }
+		public decimal LoanAmount { get; set; }
 		public decimal DefaultRate { get; set; }
 		public decimal InitiationFee { get; set; }
 
@@ -13,29 +13,29 @@
 		public decimal TenureAsPercentOfLoanTerm { get; set; }
 		public int TenureAsMonthsOfLoanTerm { get; set; }
 		public decimal CollectionRate { get; set; }
-		public int Cogs { get; set; }
+		public decimal Cogs { get; set; }
 		public decimal DebtPercentOfCapital { get; set; }
 		public decimal CostOfDebt { get; set; }
-		public int OpexAndCapex { get; set; }
-		public int ProfitBeforeTax { get; set; }
+		public decimal OpexAndCapex { get; set; }
+		public decimal ProfitBeforeTax { get; set; }
 
 		// Main output fields
 		public decimal MonthlyInterestToCharge { get; set; }
-		public int SetupFeeForEuLoan { get; set; }
+		public decimal SetupFeeForEuLoan { get; set; }
 		public decimal EuLoanPercentages { get; set; } // 1.75 or 2
 
 		// Other output fields
 		public decimal InterestRate { get; set; }
-		public int AverageLoanAmount { get; set; }
-		public int AverageRevenuePerLoan { get; set; }
-		public int CogsOutput { get; set; }
-		public int GrossProfit { get; set; }
-		public int OpexAndCapexOutput { get; set; }
-		public int Ebitda { get; set; }
-		public int NetLossFromDefaults { get; set; }
-		public int CostOfDebtOutput { get; set; }
-		public int TotalCost { get; set; }
-		public int ProfitBeforeTaxOutput { get; set; }
-		public int Balance { get; set; }
+		public decimal AverageLoanAmount { get; set; }
+		public decimal AverageRevenuePerLoan { get; set; }
+		public decimal CogsOutput { get; set; }
+		public decimal GrossProfit { get; set; }
+		public decimal OpexAndCapexOutput { get; set; }
+		public decimal Ebitda { get; set; }
+		public decimal NetLossFromDefaults { get; set; }
+		public decimal CostOfDebtOutput { get; set; }
+		public decimal TotalCost { get; set; }
+		public decimal ProfitBeforeTaxOutput { get; set; }
+		public decimal Balance { get; set; }
 	}
 }
